Refactor: move legend-item into legend-view
